+++ example_output/_templates/test-generator-doubled-with-sfx/new/test_strings.json.ejs.t
@@ -1,5 +1,5 @@
 ---
-to: <%= name %>/test_strings.json
+to: test_strings.json
 ---
 {
     "about this file": [