+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+- add: option to specify whether generator will create a parent dir (`hygen-create setopt --gen-parent-dir|--no-parent-dir`)
+- change: the created generator will not include a parent dir by default
 - fix: HYGEN_CREATE_TMPLS can be relative
 
 ## [0.1.1] - 2017-02-28